+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
 ## [Unreleased]
+
+## [1.7.0] - 2019-03-23
+### Added
+- `printHelpOnEmptyArgs` parameter to `CliktCommand` constructor. ([#41](https://github.com/ajalt/clikt/issues/41))
+
 ### Fixed
-- Arguments with `multiple(required=true)` now report an error if no argument is given on the command line.
+- Usage errors now correctly print subcommand names. ([#47](https://github.com/ajalt/clikt/issues/47))
+- Arguments with `multiple(required=true)` now report an error if no argument is given on the command line. ([#36](https://github.com/ajalt/clikt/issues/36))
 
 ## [1.6.0] - 2018-12-02
 ### Added